Tidy up paths. Make build work in Xcode.
- all build stuff goes in build/ rather than 0.*/
- fix a few Xcode warnings
- change `cmake' to `${CMAKE_COMMAND' as necessary
- update docs
- poke release script (untested)

+++ src/b2/misc.cpp
@@ -499,9 +499,9 @@ decode(uint32_t* state,uint32_t* codep,uint32_t byte) {
 
 static int GetBBCChar(uint32_t codepoint) {
     if(codepoint==13||codepoint==10) {
-        return codepoint;
+        return (int)codepoint;
     } else if(codepoint>=32&&codepoint<=126) {
-        return codepoint;
+        return (int)codepoint;
     } else if(codepoint==0xa3) {
         return 95;//GBP symbol
     } else {
